html/canvas/offscreen/layers/2d.layer.putImageData.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.layer.putImageData</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.putImageData</h1>
<p class="desc">Check that calling putImageData in a layer throws an exception.</p>


<script>
var t = async_test("Check that calling putImageData in a layer throws an exception.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.beginLayer();
  const canvas2 = new OffscreenCanvas(100, 50);
  const ctx2 = canvas2.getContext('2d')
  const data = ctx2.getImageData(0, 0, 1, 1);
  assert_throws_dom("InvalidStateError", () => ctx.putImageData(data, 0, 0));
  t.done();

});
</script>
